+++ src/Connection.php
@@ -7,8 +7,9 @@ namespace Yiisoft\Db\Pgsql;
 use Yiisoft\Db\Driver\Pdo\AbstractPdoConnection;
 use Yiisoft\Db\Driver\Pdo\PdoCommandInterface;
 use Yiisoft\Db\Exception\InvalidArgumentException;
-use Yiisoft\Db\Pgsql\Column\ColumnBuilder;
+use Yiisoft\Db\Pgsql\Column\ColumnFactory;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
+use Yiisoft\Db\Schema\Column\ColumnFactoryInterface;
 use Yiisoft\Db\Schema\Quoter;
 use Yiisoft\Db\Schema\QuoterInterface;
 use Yiisoft\Db\Schema\SchemaInterface;
@@ -45,9 +46,9 @@ final class Connection extends AbstractPdoConnection
         return new Transaction($this);
     }
 
-    public function getColumnBuilderClass(): string
+    public function getColumnFactory(): ColumnFactoryInterface
     {
-        return ColumnBuilder::class;
+        return new ColumnFactory();
     }
 
     public function getLastInsertID(string $sequenceName = null): string